+++ src/data/BlockModelDefinition.go
@@ -10,28 +10,28 @@ type BlockModelDefinition struct {
 }
 
 type TransformDefinition struct {
-	Rotation    *[]float32 `json:"rotation,omitempty"`
-	Translation *[]float32 `json:"translation,omitempty"`
-	Scale       *[]float32 `json:"scale,omitempty"`
+	Rotation    *[]float64 `json:"rotation,omitempty"`
+	Translation *[]float64 `json:"translation,omitempty"`
+	Scale       *[]float64 `json:"scale,omitempty"`
 }
 
 type ElementDefinition struct {
-	From     []float32                  `json:"from,omitempty"`
-	To       []float32                  `json:"to,omitempty"`
+	From     []float64                  `json:"from,omitempty"`
+	To       []float64                  `json:"to,omitempty"`
 	Rotation *ElementRotationDefinition `json:"rotation,omitempty"`
 	Faces    map[string]FaceDefinition  `json:"faces,omitempty"`
 	Shade    *bool                      `json:"shade,omitempty"`
 }
 
 type ElementRotationDefinition struct {
-	Angle   float32   `json:"angle"`
+	Angle   float64   `json:"angle"`
 	Axis    string    `json:"axis"`
-	Origin  []float32 `json:"origin,omitempty"`
+	Origin  []float64 `json:"origin,omitempty"`
 	Rescale *bool     `json:"rescale,omitempty"`
 }
 
 type FaceDefinition struct {
-	Uv        []float32 `json:"uv,omitempty"`
+	Uv        []float64 `json:"uv,omitempty"`
 	Texture   string    `json:"texture"`
 	Rotation  *int      `json:"rotation,omitempty"`
 	TintIndex *int      `json:"tintindex,omitempty"`
